requirements.txt
openai
streamlit
pandas
plotly
duckdb
python-dotenv